Performance: remove some 'require' inside methods
MRI 1.8.7/1.9.3 do a full search in the full load path at every call
to require, even though nothing will be loaded at the end.
This can be slow if require is called in a method that is called
a lot.

This saves a few percents on compilation time on my setup.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>