Performance: negatively cache not loaded types per compilation
When checking a given resource_type, we defer to the builtin type
system. This type system tries to autoload the type as ruby classes
in the ruby loadpath.
If the resource in question happens to be a definition that might be
used a lot of time we would constantly try to load it as a builtin
type from disk.

This patch adds a per thread negative cache system to the type manager.

Note: we could also have checked for definition in P::Resource#resource_type
before going to ruby types, but that would change the semantic of
load ordering. Because at initialization, the master uses some built-in
types to setup dependent files, those would be checked as definitions
first, and would trigger a call to known_resource_type which would
then load the whole user manifests way too early. Thus adding
negative caching seems the sanest solution :)

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>